+++ src/Service/CloudWatchLogs/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - AWS enhancement: Documentation updates.
+- AWS api-change: Bug fix: logGroupName is now not a required field in GetLogEvents, FilterLogEvents, GetLogGroupFields, and DescribeLogStreams APIs as logGroupIdentifier can be provided instead
 
 ## 1.5.0